meshbox-os/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meshbox"
version = "5.0.0"
description = "Decentralized encrypted mesh network with SANP protocol over Tor"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "MeshBox Team"},
]
keywords = ["mesh", "encrypted", "decentralized", "communication", "offline", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "PyNaCl>=1.5.0",
    "click>=8.1.0",
    "msgpack>=1.0.0",
]

[project.optional-dependencies]
web = [
    "Flask>=3.0.0",
]
bluetooth = [
    "bleak>=0.21.0",
]
qr = [
    "qrcode>=7.4.0",
    "Pillow>=10.0.0",
]
network = [
    "zeroconf>=0.131.0",
    "aiohttp>=3.9.0",
]
tor = [
    "stem>=1.8.0",
    "PySocks>=1.7.0",
]
sanp = [
    "msgpack>=1.0.0",
    "stem>=1.8.0",
    "PySocks>=1.7.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
]
all = [
    "Flask>=3.0.0",
    "bleak>=0.21.0",
    "zeroconf>=0.131.0",
    "aiohttp>=3.9.0",
    "qrcode>=7.4.0",
    "Pillow>=10.0.0",
    "cryptography>=42.0.0",
    "stem>=1.8.0",
    "PySocks>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/meshbox/meshbox"
Repository = "https://github.com/meshbox/meshbox"
Issues = "https://github.com/meshbox/meshbox/issues"

[project.scripts]
meshbox = "meshbox.cli:main"

[tool.setuptools.packages.find]
include = ["meshbox*"]

[tool.setuptools.package-data]
"meshbox.web" = [
    "templates/**/*.html",
    "static/**/*",
]
